fix: rename CDM to CMD
- use copy hook instead of function
