+++ pages/utilities/base-64-encoder.tsx
@@ -6,13 +6,14 @@ import { Card } from "../../components/ds/CardComponent";
 import { Button } from "../../components/ds/ButtonComponent";
 import { Label } from "../../components/ds/LabelComponent";
 import Header from "../../components/Header";
-import { CMDK } from "../../components/CDMK";
+import { CMDK } from "../../components/CMDK";
+import { useCopyToClipboard } from "../../components/hooks/useCopyToClipboard";
 
 export default function Base64Encoder() {
   const [type, setType] = useState<"encoder" | "decoder">("encoder");
   const [input, setInput] = useState("");
   const [output, setOutput] = useState("");
-  const [buttonText, setButtonText] = useState("Copy");
+  const { buttonText, handleCopy } = useCopyToClipboard();
 
   const handleChange = useCallback(
     (event: React.ChangeEvent<HTMLTextAreaElement>) => {
@@ -27,13 +28,6 @@ export default function Base64Encoder() {
     },
     [type]
   );
-
-  const handleCopy = useCallback(() => {
-    navigator.clipboard.writeText(output).then(() => {
-      setButtonText("Copied!");
-      setTimeout(() => setButtonText("Copy"), 1200);
-    });
-  }, [output]);
 
   const setActiveTab = (type: "encoder" | "decoder") => {
     setType(type);
@@ -86,7 +80,7 @@ export default function Base64Encoder() {
             />
             <Label>{type === "encoder" ? "Base64" : "Text"}</Label>
             <Textarea value={output} rows={6} readOnly className="mb-4" />
-            <Button variant="outline" onClick={handleCopy}>
+            <Button variant="outline" onClick={() => handleCopy(output)}>
               {buttonText}
             </Button>
           </div>
