+++ pages/utilities/yaml-to-json.tsx
@@ -7,7 +7,7 @@ import { Label } from "../../components/ds/LabelComponent";
 import Header from "../../components/Header";
 import { useCopyToClipboard } from "../../components/hooks/useCopyToClipboard";
 import yaml from "js-yaml";
-import { CMDK } from "../../components/CDMK";
+import { CMDK } from "../../components/CMDK";
 
 export default function YAMLtoJSON() {
   const [input, setInput] = useState("");
